requirements.txt
requests
BeuatifulSoup4
win11toast
groq
python-dotenv